pyproject.toml
[tool.poetry]
name = "support"
version = "0.1.0"
description = ""
authors = ["Televizor1945 <kurochkin.egor1993@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.4"
djangorestframework = "^3.14.0"
djangorestframework-jwt = "^1.11.0"
psycopg2 = "^2.9.5"
django-filter = "^22.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
